+++ apps/cloud/src/app/features/xpert/studio/components/context-menu/context-menu.component.html
@@ -98,7 +98,7 @@
 
 <ng-template #blockMenu>
   <div cdkMenu class="cdk-menu__medium bg-components-panel-bg-blur rounded-lg backdrop-blur-[5px] shadow-lg !min-w-[256px]">
-    <div class="flex items-start px-2 mt-2 text-sm font-medium text-text-tertiary">
+    <div class="flex items-start px-2 mt-2 text-base font-medium text-text-tertiary">
       {{ 'PAC.Xpert.Logic' | translate: {Default: 'Logic'} }}
     </div>
     <div cdkMenuItem class="flex items-center px-4 w-full h-9 rounded-lg hover:bg-state-base-hover cursor-pointer"
@@ -111,7 +111,7 @@
           <g id="icons/if-else"><path id="Vector (Stroke)" fill-rule="evenodd" clip-rule="evenodd" d="M8.16667 2.98975C7.80423 2.98975 7.51042 2.69593 7.51042 2.3335C7.51042 1.97106 7.80423 1.67725 8.16667 1.67725H11.0833C11.4458 1.67725 11.7396 1.97106 11.7396 2.3335V5.25016C11.7396 5.6126 11.4458 5.90641 11.0833 5.90641C10.7209 5.90641 10.4271 5.6126 10.4271 5.25016V3.91782L7.34474 7.00016L10.4271 10.0825V8.75016C10.4271 8.38773 10.7209 8.09391 11.0833 8.09391C11.4458 8.09391 11.7396 8.38773 11.7396 8.75016V11.6668C11.7396 12.0293 11.4458 12.3231 11.0833 12.3231H8.16667C7.80423 12.3231 7.51042 12.0293 7.51042 11.6668C7.51042 11.3044 7.80423 11.0106 8.16667 11.0106H9.49901L6.14484 7.65641H1.75C1.38756 7.65641 1.09375 7.3626 1.09375 7.00016C1.09375 6.63773 1.38756 6.34391 1.75 6.34391H6.14484L9.49901 2.98975H8.16667Z" fill="currentColor"></path></g>
         </svg>
       </div>
-      <div class="text-sm text-text-secondary">
+      <div class="text-text-secondary">
         {{ 'PAC.Xpert.Router' | translate: {Default: 'Router'} }}
       </div>
     </div>
@@ -134,7 +134,7 @@
         shrink-0 mr-2">
         <i class="ri-repeat-2-line text-sm"></i>
       </div>
-      <div class="text-sm text-text-secondary">{{'PAC.Xpert.Iterating' | translate: {Default: 'Iterating'} }}</div>
+      <div class="text-text-secondary">{{'PAC.Xpert.Iterating' | translate: {Default: 'Iterating'} }}</div>
     </div>
 
     <div cdkMenuItem class="flex items-center px-4 w-full h-9 rounded-lg hover:bg-state-base-hover cursor-pointer"
@@ -145,10 +145,10 @@
         shrink-0 mr-2">
         <i class="ri-chat-ai-fill text-sm"></i>
       </div>
-      <div class="text-sm text-text-secondary">{{'PAC.Xpert.Answer' | translate: {Default: 'Answer'} }}</div>
+      <div class="text-text-secondary">{{'PAC.Xpert.Answer' | translate: {Default: 'Answer'} }}</div>
     </div>
 
-    <div class="flex items-start px-2 mt-2 text-sm font-medium text-text-tertiary">
+    <div class="flex items-start px-2 mt-2 font-medium text-text-tertiary">
       {{ 'PAC.Xpert.Others' | translate: {Default: 'Others'} }}
     </div>
     <div cdkMenuItem class="ngm-cdk-menu-item shrink-0 px-4 w-full h-9"
@@ -158,7 +158,7 @@
         mr-2 shrink-0">
         <i class="ri-sticky-note-add-line"></i>
       </div>
-      <div class="text-sm text-text-secondary">
+      <div class="text-text-secondary">
         {{ 'PAC.Xpert.Note' | translate: {Default: 'Note'} }}
       </div>
     </div>
