+++ apps/cloud/src/app/features/xpert/studio/components/toolset-menu/toolset.component.html
@@ -44,8 +44,8 @@
       #overlayTrigger="cdkOverlayOrigin"
       >
       <emoji-avatar [avatar]="toolset.avatar" xs class="shrink-0 mr-2 overflow-hidden rounded-full scale-95"  />
-      <div class="flex-1 truncate" [ngmHighlight]="searchText()" [content]="toolset.name">
-        {{toolset.name}}
+      <div class="flex-1 truncate" [ngmHighlight]="searchText()" [content]="toolset.label | i18n">
+        {{toolset.label | i18n}}
       </div>
     </div>
   }
